Move code from main Gui to Controller
